factor tree_view js into functions, form for picking root_table/cond
setupTree can now be called on AJAX callback
so we can e.g. choose a bigger height if there's a lot of nodes

however, getting a funny glitch when I first render the tree.
the root connection starts at the top of the screen, fragmented
and piles itself down to the right place.

probably a variable scoping issue that came up in factoring.

rename tree_table to root_table, since the relationships are going to
allow additional tables to be joined into the tree