tree - UNDER CONSTRUCTION - adding vals to ALL relationships
I was having a problem where a child was related to a parent by relationship1
and then the child has children too but was not being considered in the query
for being that kind of parent.

I realized that the parent_val arrays were shoehorning nodes into
needing the same relationship that they themselves were discovered by.

Still haven't totally fixed it but getting closer.