(UNDER CONSTRUCTION) obj_editor: option to dump stored rows as JSON
Sometimes when I save the stored_rows from my phone to the DB
some of them fail to be INSERTed because of a nonexisted
table or field

Then it's really awkward to get them back, because the local
browser session doesn't realize anything went wrong, and
the only way to try again is the Recover All Stored Rows,
and then I get tons of duplicates.

Instead, I'm going to dump a JSON file of all the rows so I
have it on the computer and can choose what to do.