change selectTable input into a real form
this is because the damn selectFirstFormField
is still not working on my phone even though it
worked before...

but apparently iOS has a "security feature" that
prevents focusing on a form field based on an arbitrary
event (or let alone no event at all).  My keypress
event is not "trusted" enough, so input.focus() silently
fails.

so this approach is to use a real <form> around the
selectTable input so that the submit event is what happens
rather than the keypress event.

but then we lose the alt-Enter / ctrl-Enter functionality,
and it seems like the focus is STILL not working, so most
likely I will revert this soon.