tree: ALMOST got array match working (pgsql only)
SQL query is rewritten in this case as
val1 = ANY(parent_field) or val2 = ANY(parent_field) ...

tree is showing up fine ...
BUT nodes still can't show up in more than one place!
it seems this is because of the master list of nodes
that prevents a copy from being made.  I guess now we
WANT copies to be made, so the nodes can have different
parents. ?