Re-tighten Db::build_where_clause (add $strict_wheres param)
In a recent commit, I added a Config that provides the possibility
of "loosening" the way the WHERE clauses are built up,
allowing for example the use of @> for Postgres Arrays instead of
a strict =.

However, making this default behavior in build_where_clause()
creates an undesirable and dangerous coupling to the global Config,
and affects multiple places where WHERE clauses are built up -
not all of which need or want this new "looser matching" functionality.

In this commit, we add a $strict_wheres parameter to build_where_clauses(),
defaulting it to true and therefore restoring the previous behavior for
all but specific calls to the function.  The Db::get() fn that is used
in crud_api (i.e. for Ctrl+S searches, View button, etc), is then altered
to set $strict_wheres=false so that we can still use the new
"looser matching" Config, while still having a tight, predictable
build_where_clauses() behavior by default.