factor out Predicate::loosely_interpret_where_clause() fn
Used to be a bunch of logic straight in Db::build_where_clause()
but took it out into its own fn, for simplicity and clarity.

To avoid performance penalty of copying objects / arrays,
the function takes $key and $predicate_or_val by reference
and makes changes as needed directly to those zvals.