fix additional issue with Ctrl+S search / View Button
a new issue came up as a result of last commit's fix

Namely, in the last commit I added a $strict_wheres check
to cause the "looser matching" logic to not be the default
behavior for Db::build_where_clause().  But when I did this
it also disabled the "looser matching" logic on e.g.
Ctrl+S searches, because those make a crud_api call to
action=view, which uses expand_tablename_into_query() instead
of Db::get().

In this commit, I fix this issue by adding the optional
$strict_wheres param to expand_tablename_into_query(), and
pass in a value for it in crud_api action=view - namely,
I pass in $strict_wheres=true iff I have the $match_aliases
option (because this was the case where the query-building was
failing).