Fix an issue with auth.php redirection and guessing the uri trunk
The Config::guess_uri_trunk() mechanism really needs more work
but this fixes one related issue that I found in the auth redirect

Some background:

For the most part, every entry point php file run within db_viewer
initially defines the vars $cur_view and $cur_subview
then when the Config system loads the configs,
it builds up the assumptions of the paths of where everything is
based on an initial calculation (a "guess") of where the $uri_trunk is.

This guess is about regex-matching the REQUEST_URI to find what prefix
is found prior to the /$cur_view/$cur_subview.php part.
This prefix is called the "uri_trunk".

The auth.php problem:

auth.php was setting cur_view to 'auth'.  In the current scheme,
this implies a uri of '/auth/index.php', because cur_view = 'auth'
and cur_subview is nothing (implying index.php).  This can lead to
incorrect redirects.

The fix was to have auth.php set cur_subview to 'auth' instead, while
allowing cur_view to be '', (indicating a root-level subview).  Therefore
the implied uri is: '/auth.php', which redirects properly.